zephyrCommon: Implement analogWrite()
The 'pwms' node under the 'zephyr,user' defines available PWM channels.

The 'pwm-pins' defines the mapping of PWM channels to pin numbers.
API looks up the 'pwm-pins' when querying the PWM channel by digital
pin number. So it should be in the same order as 'pwms' node.

PWM channels should configure and tied to the output pins.
The period property of the PWM channel is used as resolution as PWM.
For original Arduino API compatibility, it should be 255.
The period property and API argument determines the duty ratio.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>