libraries: Wire: Instantiate Wire following dts configuration
Instaintiate Wire with i2c device that defined in 'i2cs' array
under zephyr,user node.
If 'i2cs' array is not defined, tries to instantiate with
i2c device that labeled as 'arduino_i2c'.

As a result of apply new rule, the i2c_dev that declared in pinmap.h
is no more used, remove it.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>