libraries: SPI: handle multiple SPI instances
Create multiple SPI instances if the `spis` array contains plural elements.
Declare these as 'SPI1', 'SPI2', ..., from the second element of the array.
(The first element is already declared as 'SPI'.)

If `spis` is not defined but the DTS already defines `arduino-spi` and,
use it to define the first SPI instance.
The `arduion-spi` is usually defined with arduino_header in
boards definitions.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>