SPI - Fix Transfer16 plus allow subclass to gain access to member variables
resolves #13

And: provides for c) option in #14

In particular, added another config option (config16) that is initialized at beginTransaction with same
settings as config except word size set to 16 instead of 8.

Also: updated begin() to call beginTransaction with default SPISettings, and the endTransaction, such that sketches that don't call beginTransaction output at the default settings.

Changed the header file private: to protected