Fix micros() for nano without breaking giga
The rollover fix for the giga used the 64 bit timer instead of 32
bits.  The nano does not have a 64 bit timer, so you need to use
the 32 bit instead.

So I know check to see if the processor supports 64 bits if so use it
else fall back to 32 bit.

Note: This branch doesn't include support for Arduino Giga,
      but we'll include it as it's a reasonable solution for
      preventing timer overflows.

Co-Authored-by: Kurt Eckhardt <kurte@rockisland.com>
Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>