If sketch is not connected to serial - dump serial output
Current code would hang after a few Serial.print like statements happened when we are not connected to
the Serial monitor.

So I put overrides in for the Serial.write(buf, cnt) and Serial.flush()
That would first check to see if Serial returns true (DTR)...

This appears to match the behavior of the MBED Portenta H8 code.

Co-authored-by: kurte <kurte@rockisland.com>
Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>