zephyrCommon: Changing pin numbering rules
Zephyr's ba48d83b changes make it easy to configure pin numbering
at compile, so we'll change the pin numbers to be based on
the GPIO pin numbers.

This is same as the traditional Arduino, using the GPIO numbers as
pin numbers. Pin names written on the board, such as D1 and D2,
are aliases for these pin numbers.

When using multiple GPIO ports, pin numbers are numbered consecutively
from the beginning. For example, if you have two 16-port GPIOs,
16 refers to the first pin (idx=0) of the second port.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>