Arduino_LED_Matrix: add const to matrixWrite
This allows you to then change the sketch(s) used within the test apps
for the Q to use this header file with little changes to them.

Like the Weather forecast on LED matrix

    #include "Arduino_LED_Matrix.h"
    Arduino_LED_Matrix matrix;

    void setup() {
      matrix.begin();
      ...

Or in my test app I also then included ArduinoGraphics

    #include <Arduino_RouterBridge.h>
    #include "ArduinoGraphics.h"
    ...

Note by committer: Only under the `loader` was picked.

Co-authored-by: Kurt Eckhardt <kurte@rockisland.com>
Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>